// src/utils/enums.ts
export enum ReportsEndpoints {
    ALUNOS = "alunos",
    ANIVERSARIOS = "aniversarios",
    ATIVIDADES_EXTRA_CLASSE = "atividades-extra-classe",
    BOLETOS = "boletos",
    CONTATO_TELEFONICO = "contato-telefonico",
    CONTATO_TELEFONICO_ANIVERSARIO = "contato-telefonico-aniversario",
    PERIODOS = "periodos",
    PERMISSAO_IMAGEM = "permissao-imagem"
}